move player instance logic to hook
